UPSTREAM: soc: qcom: ipa: IPA endpoints
This patch includes the code implementing an IPA endpoint.  This is
the primary abstraction implemented by the IPA.  An endpoint is one
end of a network connection between two entities physically
connected to the IPA.  Specifically, the AP and the modem implement
endpoints, and an (AP endpoint, modem endpoint) pair implements the
transfer of network data in one direction between the AP and modem.

Endpoints are built on top of GSI channels, but IPA endpoints
represent the higher-level functionality that the IPA provides.
Data can be sent through a GSI channel, but it is the IPA endpoint
that represents what is on the "other end" to receive that data.
Other functionality, including aggregation, checksum offload and
(at some future date) IP routing and filtering are all associated
with the IPA endpoint.

Signed-off-by: Alex Elder <elder@linaro.org>
Signed-off-by: David S. Miller <davem@davemloft.net>
(cherry picked from commit 84f9bd12d46dbe8ac7d4b650a8fbb4c49657a38b)

BUG=b:74020136
TEST=Boot modem on Trogdor

Signed-off-by: Evan Green <evgreen@chromium.org>
Change-Id: I391d8a41176c344301f484ee51dc357691b790d2
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/kernel/+/2086674
Reviewed-by: Sean Paul <seanpaul@chromium.org>
Reviewed-by: Matthias Kaehlcke <mka@chromium.org>