UPSTREAM: soc: qcom: ipa: filter and routing tables
This patch contains code implementing filter and routing tables for
the IPA.  A filter table allows rules to be used for filtering
packets that depart the AP at an endpoint.  A filter table entry
contains the address of a set of rules to apply for each endpoint
that supports filtering.

A routing table allows packets to be routed to an endpoint based
on packet metadata.  It is also a table whose entries each contain
the address of a set of routing rules to apply.

Neither filtering nor routing is supported by the current driver.
All table entries refer to rules that mean "no filtering" and "no
routing."

Signed-off-by: Alex Elder <elder@linaro.org>
Signed-off-by: David S. Miller <davem@davemloft.net>
(cherry picked from commit 2b9feef2b6c288c1f11127fe236bb5078d5d51b3)

BUG=b:74020136
TEST=Boot modem on Trogdor

Signed-off-by: Evan Green <evgreen@chromium.org>
Change-Id: I9b2fb3121f74428e83874e71a653f28bc267eada
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/kernel/+/2086675
Reviewed-by: Sean Paul <seanpaul@chromium.org>
Reviewed-by: Matthias Kaehlcke <mka@chromium.org>