UPSTREAM: drm/i915: Improve the start alignment of bonded pairs
Always wait on the start of the signaler request to reduce the problem
of dequeueing the bonded pair too early -- we want both payloads to
start at the same time, with no latency, and yet still allow others to
make full use of the slack in the system. This reduce the amount of time
we spend waiting on the semaphore used to synchronise the start of the
bonded payload.

Signed-off-by: Chris Wilson <chris@chris-wilson.co.uk>
Reviewed-by: Tvrtko Ursulin <tvrtko.ursulin@intel.com>
Link: https://patchwork.freedesktop.org/patch/msgid/20200306133852.3420322-3-chris@chris-wilson.co.uk
(cherry picked from commit 798fa870abf519e8269bb63536d31db773a860d9)

BUG=b:175077993
TEST=Tested on eve-kernelnext

Signed-off-by: Sean Paul <seanpaul@chromium.org>
Change-Id: Ic60b4281e96b864f2996e845a309a2d897f5881b