UPSTREAM: xhci: Finetune host initiated USB3 rootport link suspend and resume
Depending on the current link state the steps to resume the link to U0
varies. The normal case when a port is suspended (U3) we set the link
to U0 and wait for a port event when U3exit completed and port moved to
U0.

If the port is in U1/U2, then no event is issued, just set link to U0

If port is in Resume or Recovery state then the device has already
initiated resume, and this host initiated resume is racing against it.
Port event handler for device initiated resume will set link to U0,
just wait for the port to reach U0 before returning.

Signed-off-by: Mathias Nyman <mathias.nyman@linux.intel.com>
Link: https://lore.kernel.org/r/20200312144517.1593-9-mathias.nyman@linux.intel.com
Signed-off-by: Greg Kroah-Hartman <gregkh@linuxfoundation.org>
(cherry picked from commit ceca49382ac20e06ce04c21279c7f2868c4ec1d4)

BUG=b:140290596
TEST=cat /sys/bus/usb/devices/*/power/runtime_status
TEST=Check before and after hotplugs that devices are
entering 'suspended' states when not in use.
Devices to check:
- USB Type A/C pen drives
- Type C hub

Signed-off-by: Casey Bowman <casey.g.bowman@intel.com>
Change-Id: I59227806b3b0150f44e5181c532eebb3026e3ef5
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/kernel/+/2103668
Reviewed-by: Sean Paul <seanpaul@chromium.org>
Reviewed-by: Prashant Malani <pmalani@chromium.org>