CHROMIUM: iwl7000: mvm: add an option to add PASN station
A FTM responder may do PASN authentication with unassociated stations
to allow secure ranging. In this case, the driver will add an internal
station and install the TK so the FW will accept protected FTM
request frames from this station and will send a protected FTM
response frame.

In addition, the driver needs to configure the HLTK to the FW so
the FW can derive the secure LTF bits. This is left for a later
patch since it is not yet supported by the FW.

Change-Id: I72ef7f9753964555561c27ec503241105eddb14e
Signed-off-by: Avraham Stern <avraham.stern@intel.com>
iwl7000-tree: 9f463264075b51b4491843c18a8f620d0e43c7d5