UPSTREAM: drm: panel: Add Starry KR070PE2T
The KR070PE2T is a 7" panel with a resolution of 800x480.

KR070PE2T is the marking present on the ribbon cable. As this panel is
probably available under different brands, this marking will catch
most devices.

As I can't find a datasheet for this panel, the bus_flags are instead
from trial-and-error. The flags seem to be common for these kind of
panels as well.

Signed-off-by: Pascal Roeleven <dev@pascalroeleven.nl>
Signed-off-by: Sam Ravnborg <sam@ravnborg.org>
Link: https://patchwork.freedesktop.org/patch/msgid/20200320112205.7100-3-dev@pascalroeleven.nl
(cherry picked from commit 105235e4ae4cbdf068116ca0526567cfae79076d)

BUG=b:179329054
TEST=Future patches pick cleanly

Signed-off-by: Douglas Anderson <dianders@chromium.org>
Change-Id: I2f3eb2d3a4499221cc7940dae9be433625109433
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/kernel/+/3227655
Reviewed-by: Sean Paul <seanpaul@chromium.org>
Reviewed-by: Stéphane Marchesin <marcheu@chromium.org>