UPSTREAM: coresight: cti: Add device tree support for v8 arch CTI
The v8 architecture defines the relationship between a PE, its optional ETM
and a CTI. Unlike non-architectural CTIs which are implementation defined,
this has a fixed set of connections which can therefore be represented as a
simple tag in the device tree.

This patch defines the tags needed to create an entry for this PE/ETM/CTI
relationship, and provides functionality to implement the connection model
in the CTI driver.

Signed-off-by: Mike Leach <mike.leach@linaro.org>
Reviewed-by: Suzuki K Poulose <suzuki.poulose@arm.com>
Signed-off-by: Mathieu Poirier <mathieu.poirier@linaro.org>
Link: https://lore.kernel.org/r/20200320165303.13681-7-mathieu.poirier@linaro.org
Signed-off-by: Greg Kroah-Hartman <gregkh@linuxfoundation.org>
(cherry picked from commit cffd054f1708a07330536a9704b70bc5d090cd98)

BUG=b:149851136
TEST=See /sys/bus/coresight has devices on Trogdor

Signed-off-by: Stephen Boyd <swboyd@chromium.org>
Change-Id: I9f68063b76f0f5e01a41a6dfa2ddccab7d7fa747
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/kernel/+/2161808
Reviewed-by: Douglas Anderson <dianders@chromium.org>
Reviewed-by: Sean Paul <seanpaul@chromium.org>