UPSTREAM: drm/i915/selftests: Measure the energy consumed while in RC6
Measure and compare the energy consumed, as reported by the rapl MSR,
by the GPU while in RC0 and RC6 states. Throw an error if RC6 does not
at least halve the energy consumption of RC0, as this more than likely
means we failed to enter RC0 correctly.

If we can't measure the energy draw with the MSR, then it will report 0
for both measurements. Since the measurement works on all gen6+, this seems
worth flagging as an error.

Signed-off-by: Chris Wilson <chris@chris-wilson.co.uk>
Cc: Mika Kuoppala <mika.kuoppala@linux.intel.com>
Cc: Andi Shyti <andi.shyti@intel.com>
Reviewed-by: Andi Shyti <andi.shyti@intel.com>
Link: https://patchwork.freedesktop.org/patch/msgid/20200325101502.12591-1-chris@chris-wilson.co.uk
(cherry picked from commit 032d992dcbefb1e6a43832fa36c0da11b6b64741)

BUG=b:168669388
TEST=Check for GPU frequency samples above 1100MHz.
     - Login as guest and launch WebGL Aquarium with 30,000 fish.
     - Monitor frequency statistics.
       `socwatch -f cpu -f gfx -t 120`
       alternatively
       `while true; do cat /sys/class/drm/card0/gt_act_freq_mhz; sleep 0.1; done`

Signed-off-by: Ed Baker <edward.baker@intel.com>
Change-Id: Ie5da47c4267baaef3cac09d888b6463525891784
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/kernel/+/2472667
Reviewed-by: Ilja H. Friedel <ihf@chromium.org>
Reviewed-by: Sean Paul <seanpaul@chromium.org>