FROMGIT: soc / drm: mediatek: Fix mediatek-drm device probing
In the actual implementation the same compatible string
"mediatek,<chip>-mmsys" is used to bind the clock drivers
(drivers/soc/mediatek) as well as to the gpu driver
(drivers/gpu/drm/mediatek/mtk_drm_drv.c). This ends with the problem
that the only probed driver is the clock driver and there is no display
at all.

In any case having the same compatible string for two drivers is not
correct and should be fixed. To fix this, and maintain backward
compatibility, we can consider that the mmsys driver is the top-level
entry point for the multimedia subsystem, so is not a pure clock
controller but a system controller, and the drm driver is instantiated
by that MMSYS driver.

Signed-off-by: Enric Balletbo i Serra <enric.balletbo@collabora.com>
Reviewed-by: CK Hu <ck.hu@mediatek.com>
Acked-by: CK Hu <ck.hu@mediatek.com>
Signed-off-by: Matthias Brugger <matthias.bgg@gmail.com>

(cherry picked from commit 1d367541adedd1d24b423454240bc6cc1333a5ef
https://git.kernel.org/pub/scm/linux/kernel/git/matthias.bgg/linux.git/ v5.6-next/soc)

BUG=b:149607803
TEST=boot hana on 5.4 with ToT stack

Signed-off-by: Hsin-Yi Wang <hsinyi@chromium.org>
Change-Id: Idff5a5a79a72e37bbd17ef9fb64e855a2a935706
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/kernel/+/2124328
Reviewed-by: Sean Paul <seanpaul@chromium.org>
Reviewed-by: Nicolas Boichat <drinkcat@chromium.org>
Commit-Queue: Nicolas Boichat <drinkcat@chromium.org>