UPSTREAM: iio: cros_ec: Report hwfifo_watermark_max
Report the maximum amount of sample the EC can hold.
This is not tunable, but can be useful for application to find out the
maximum amount of time it can sleep when hwfifo_timeout is set to a
large number.

BUG=chromium:946178
TEST=tast run <ip> hardware.SensorRing with sensor ring enabled.
     iio-info with sensor ring disabled.

Reviewed-by: Jonathan Cameron <Jonathan.Cameron@huawei.com>
Signed-off-by: Enric Balletbo i Serra <enric.balletbo@collabora.com>
(cherry picked from commit cb87556068146de5c9933397706d3bde88b4a14d)
Signed-off-by: Gwendal Grignou <gwendal@chromium.org>
Change-Id: I59db89607577113b8e314931de691c8847953ef6
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/kernel/+/2223405
Reviewed-by: Sean Paul <seanpaul@chromium.org>