FROMGIT: thunderbolt: Set port configured for both ends of the link
Both ends of the link needs to have this set. Otherwise the link is not
re-established properly after sleep. Now since it is possible to have
mixed USB4 and Thunderbolt 1, 2 and 3 devices we need to split the link
configuration functionality to happen per port so we can pick the
correct implementation.

Signed-off-by: Mika Westerberg <mika.westerberg@linux.intel.com>
(cherry picked from commit e28178bf566cf7281b513856aa71a8d41aa81154
https://git.kernel.org/pub/scm/linux/kernel/git/westeri/thunderbolt.git next)

BUG=b:162700590
TEST=No regressions for the following use cases:
- S0iX NDA/DA
- TBT NVMe hotplug
- TBT dock hotplug
- TBT dock w/ USB pen drive
- TBT dock w/ display

Signed-off-by: Casey Bowman <casey.g.bowman@intel.com>
Change-Id: If4046f18ea51681f5b2fe0434d63fd46260db26f